requirements.txt
pandas
numpy
scikit-learn
mlflow
fastapi
uvicorn
python-dotenv
pyyaml
joblib  
plotly
pandas

streamlit
matplotlib
shap
python-multipart
psycopg2-binary
sqlalchemy
reportlab